Switch to `pre_get_avatar_data` filter.
In fact, we only need to filter the url of the avatar, not the avatar
img element. So instead of hook to `get_avatar` filter and override the
`get_avatar` function, we hook to `pre_get_avatar_data` filter to
manipulate the avatar url and let the core handle the reset. This reduce
duplication code and effort to follow core changes.